WIP #2: Remove useless code
